% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_method_functions.R
\name{bin_auc}
\alias{bin_auc}
\title{Calculate AUC for single Se/Sp pair}
\usage{
bin_auc(se, sp)
}
\arguments{
\item{se}{Sensitivity}

\item{sp}{Specificity}
}
\value{
Area under ROC curve
}
\description{
Calculate AUC
}
