% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{boot_ML}
\alias{boot_ML}
\title{Bootstrap Accuracy Statistic Estimation for Multimethod Data}
\usage{
boot_ML(
  type = c("binary", "ordinal", "continuous"),
  data,
  n_boot = 100,
  max_iter = 1000,
  tol = 1e-07,
  seed = NULL
)
}
\arguments{
\item{n_boot}{number of bootstrap estimates to compute}

\item{seed}{optional seed for RNG}

\item{...}{
  Arguments passed on to \code{\link[=estimate_ML]{estimate_ML}}
  \describe{
    \item{\code{type}}{A string specifying the data type of the methods under evaluation}
    \item{\code{data}}{An n_obs by n_method matrix containing the observed values for each method. If dimensions are named, row names will be used for obs_names, and column names will be used for method_names}
    \item{\code{init}}{An optional list of initial values used to seed the EM algorithm. If initial values are not provided, a function will be called on the data to estimate starting values.}
    \item{\code{max_iter}}{The maximum number of EM algorithm iterations to compute before reporting a result.}
    \item{\code{tol}}{The minimum change in statistic estimates needed to continue iterating the EM algorithm.}
    \item{\code{save_progress}}{A logical indication whether to save interim calculations used in the EM algorithm.}
  }}
}
\value{
a list containing accuracy estimates \code{v} and the parameters used
}
\description{
Bootstrap Accuracy Statistic Estimation for Multimethod Data
}
