% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{censor_data}
\alias{censor_data}
\title{Censor data randomly rowwise}
\usage{
censor_data(
  n_obs = dis$n_obs,
  first_reads_all = first_reads_all,
  n_method_subset = n_method_subset,
  n_method = n_method
)
}
\arguments{
\item{n_obs}{An integer representing the number of observations to simulate.}

\item{first_reads_all}{Used for binary methods. A logical which forces method
1 to have a result for every observation}

\item{n_method_subset}{Used for binary methods. An integer defining how
many methods to select at random to produce a result for each observation}

\item{n_method}{An integer representing the number of methods to simulate.}
}
\description{
Censor data randomly rowwise
}
