% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{define_disease_state}
\alias{define_disease_state}
\title{Define the True disease state of a simulated sample}
\usage{
define_disease_state(D = NULL, n_obs = NULL, prev = NULL)
}
\arguments{
\item{D}{True disease state}

\item{n_obs}{integer number of observations}

\item{prev}{prevalence of positive observations in population}
}
\value{
A list of features defining the true disease status of each observation
}
\description{
Define the True disease state of a simulated sample
}
