% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dmvnorm}
\alias{dmvnorm}
\title{Multivariate Normal Densities}
\usage{
dmvnorm(x, mu, sigma)
}
\arguments{
\item{x}{matrix of observations}

\item{mu}{vector of method means}

\item{sigma}{method covariance matrix}
}
\description{
Return the density of a point in a multivariate normal distribution
}
