% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_method_functions.R
\name{estimate_EM_continuous}
\alias{estimate_EM_continuous}
\title{Title}
\usage{
estimate_EM_continuous(
  data,
  init = list(prev_1 = NULL, mu_i1_1 = NULL, sigma_i1_1 = NULL, mu_i0_1 = NULL,
    sigma_i0_1 = NULL),
  max_iter = 100,
  tol = 1e-07,
  save_progress = FALSE
)
}
\arguments{
\item{data}{A matrix with observations as rows and methods as columns.}

\item{init}{A named list of initial values for prevalence, location, and dispersion of disease groups.}

\item{max_iter}{Maximum number of iterations to stop seeking convergence of l estimate.}

\item{tol}{Change in l below which l is said to converge.}

\item{save_progress}{Logical value denoting whether to output the calculated values of each iteration.}
}
\description{
Title
}
