% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{estimate_ML}
\alias{estimate_ML}
\title{Title}
\usage{
estimate_ML(
  type = c("binary", "ordinal", "continuous"),
  data,
  init = list(),
  max_iter = 1000,
  tol = 1e-07,
  save_progress = FALSE,
  ...
)
}
\arguments{
\item{type}{A string specifying the data type of the methods under evaluation}

\item{data}{An n_obs by n_method matrix containing the observed values for each method}

\item{init}{A list of initial values used in the EM algorithm}

\item{max_iter}{The maximum number of EM algorithm iterations to compute before reporting a result}

\item{tol}{The minimum change in statistic estimates needed to continue iterating the EM algorithm}

\item{save_progress}{A logical indication whether to save interim calculations used in the EM algorithm}

\item{...}{}
}
\description{
Title
}
