% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_method_functions.R
\name{estimate_ML_binary}
\alias{estimate_ML_binary}
\title{Estimate accuracy statistics and prevalence by ML}
\usage{
estimate_ML_binary(
  data,
  init = list(prev_1 = NULL, se_1 = NULL, sp_1 = NULL),
  max_iter = 100,
  tol = 1e-07,
  save_progress = TRUE
)
}
\arguments{
\item{data}{A matrix with n_obs rows and n_method columns}

\item{init}{A named list of initial values for EM algorithm}

\item{max_iter}{An integer representing the maximum number of iterations to allow before stopping}

\item{tol}{The minimum change between iterations required to continue iterating}

\item{save_progress}{A logical value indicating whether to save statistic values for each iteration.}
}
\value{
A list of statistics resulting from the application of the EM algorithm. If save_progress is TRUE, a list of each value needed by the algorithm for each iteration will also be included.
}
\description{
Estimate accuracy statistics and prevalence by ML
}
