% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_method_functions.R
\name{estimate_ML_ordinal}
\alias{estimate_ML_ordinal}
\title{Title}
\usage{
estimate_ML_ordinal(
  data,
  init = list(pi_1_1 = NULL, phi_1ij_1 = NULL, phi_0ij_1 = NULL, n_level = NULL),
  level_names = NULL,
  max_iter = 1000,
  tol = 1e-07,
  save_progress = FALSE
)
}
\arguments{
\item{save_progress}{}
}
\description{
Title
}
