% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_method_functions.R
\name{gen_multi_con}
\alias{gen_multi_con}
\title{Title}
\usage{
gen_multi_con(
  n_method = 2,
  n_obs = 100,
  prev = 0.5,
  D = NULL,
  mu_i1 = rep(12, n_method),
  sigma_i1 = diag(n_method),
  mu_i0 = rep(10, n_method),
  sigma_i0 = diag(n_method),
  method_names = NULL,
  obs_names = NULL
)
}
\arguments{
\item{n_method}{number of methods to generate simulated data for}

\item{n_obs}{number of observations to simulate}

\item{prev}{true prevalence of positives in target population}

\item{D}{binary vector representing the true disease status of the observations}

\item{mu_i1}{vector of method mean values for positive observations}

\item{sigma_i1}{covariance matrix of method positive observations}

\item{mu_i0}{vector of method mean values for negative observations}

\item{sigma_i0}{covariance matrix of method negative observations}

\item{obs_names}{vector of names of each observation}
}
\value{
list containing generated data and input parameters
}
\description{
Title
}
