% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_method_functions.R
\name{generate_multimethod_binary}
\alias{generate_multimethod_binary}
\title{Generate data for multiple binary measurement methods}
\usage{
generate_multimethod_binary(
  n_method = 3,
  n_obs = 100,
  prev = 0.5,
  D = NULL,
  se = rep(0.9, n_method),
  sp = rep(0.9, n_method),
  method_names = NULL,
  obs_names = NULL,
  n_method_subset = n_method,
  first_reads_all = FALSE
)
}
\arguments{
\item{n_method}{An integer representing the number of methods to simulate}

\item{n_obs}{An integer representing the number of observations to simulate}

\item{prev}{Optional value between 0-1 which represents the proportion of positive results in the target population}

\item{D}{Optional binary vector representing the true classification of each observation}

\item{se, sp}{Used for binary methods. A vector of length n_method of values between 0-1 representing the sensitivity and specificity of the methods}

\item{method_names}{Optional vector of names used to identify each method}

\item{obs_names}{Optional vector of names used to identify each observation}

\item{n_method_subset}{Used for binary methods. An integer defining how many methods to select at random to produce a result for each observation}

\item{first_reads_all}{Used for binary methods. A logical which forces method 1 to have a result for every observation}
}
\value{
A list containing the simulated data set and the parameters used to generate it
}
\description{
Generate data for multiple binary measurement methods
}
