% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_method_functions.R
\name{generate_multimethod_continuous}
\alias{generate_multimethod_continuous}
\title{Generate data set for multiple continuous methods measuring the same binary variable}
\usage{
generate_multimethod_continuous(
  n_method = 2,
  n_obs = 100,
  prev = 0.5,
  D = NULL,
  mu_i1 = rep(12, n_method),
  sigma_i1 = diag(n_method),
  mu_i0 = rep(10, n_method),
  sigma_i0 = diag(n_method),
  method_names = NULL,
  obs_names = NULL
)
}
\arguments{
\item{n_method}{An integer representing the number of methods to simulate}

\item{n_obs}{An integer representing the number of observations to simulate}

\item{prev}{Optional value between 0-1 which represents the proportion of positive results in the target population}

\item{D}{Optional binary vector representing the true classification of each observation}

\item{mu_i1, mu_i0}{Used for continuous methods. Vectors of length n_method of the method mean values for positive (negative) observations}

\item{sigma_i1, sigma_i0}{Used for continuous methods. Covariance matrices of method positive (negative) observations}

\item{method_names}{Optional vector of names used to identify each method}

\item{obs_names}{Optional vector of names used to identify each observation}
}
\value{
list containing generated data and input parameters
}
\description{
Generate data set for multiple continuous methods measuring the same binary variable
}
