% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{generate_multimethod_data}
\alias{generate_multimethod_data}
\title{Create data sets which simulate paired measurements of multiple methods}
\usage{
generate_multimethod_data(
  type = c("binary", "ordinal", "continuous"),
  n_method = 3,
  n_obs = 100,
  prev = 0.5,
  D = NULL,
  method_names = NULL,
  obs_names = NULL,
  ...
)
}
\arguments{
\item{type}{A string specifying the data type of the methods under evaluation}

\item{n_method}{An integer representing the number of methods to simulate}

\item{n_obs}{An integer representing the number of observations to simulate}

\item{prev}{Optional value between 0-1 which represents the proportion of positive results in the target population}

\item{D}{Optional binary vector representing the true classification of each observation}

\item{method_names}{Optional vector of names used to identify each method}

\item{obs_names}{Optional vector of names used to identify each observation}

\item{...}{
  Arguments passed on to \code{\link[=generate_multimethod_binary]{generate_multimethod_binary}}, \code{\link[=generate_multimethod_ordinal]{generate_multimethod_ordinal}}, \code{\link[=generate_multimethod_continuous]{generate_multimethod_continuous}}
  \describe{
    \item{\code{se,sp}}{Used for binary methods. A vector of length n_method of values between 0-1 representing the sensitivity and specificity of the methods}
    \item{\code{n_method_subset}}{Used for binary methods. An integer defining how many methods to select at random to produce a result for each observation}
    \item{\code{first_reads_all}}{Used for binary methods. A logical which forces method 1 to have a result for every observation}
    \item{\code{n_level}}{Used for ordinal methods. An integer representing the number of ordinal levels each method has}
    \item{\code{pmf_pos,pmf_neg}}{Used for ordinal methods. A n_method by n_level matrix representing the probability mass functions for positive and negative results, respectively}
    \item{\code{level_names}}{Used for ordinal methods. Optional vector of names used to identify each level}
    \item{\code{mu_i1,mu_i0}}{Used for continuous methods. Vectors of length n_method of the method mean values for positive (negative) observations}
    \item{\code{sigma_i1,sigma_i0}}{Used for continuous methods. Covariance matrices of method positive (negative) observations}
  }}
}
\value{
A list containing a simulated data set and the parameters used to create it
}
\description{
Create data sets which simulate paired measurements of multiple methods
}
