% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_method_functions.R
\name{generate_multimethod_ordinal}
\alias{generate_multimethod_ordinal}
\title{Generate data for multiple ordinal measurement methods}
\usage{
generate_multimethod_ordinal(
  n_method = 3,
  n_obs = 100,
  prev = 0.5,
  D = NULL,
  n_level = 5,
  pmf_pos = matrix(rep(1:n_level - 1, n_method), nrow = n_method, byrow = TRUE),
  pmf_neg = matrix(rep(n_level:1 - 1, n_method), nrow = n_method, byrow = TRUE),
  method_names = NULL,
  level_names = NULL,
  obs_names = NULL
)
}
\arguments{
\item{n_method}{An integer representing the number of methods to simulate}

\item{n_obs}{An integer representing the number of observations to simulate}

\item{prev}{Optional value between 0-1 which represents the proportion of positive results in the target population}

\item{D}{Optional binary vector representing the true classification of each observation}

\item{n_level}{Used for ordinal methods. An integer representing the number of ordinal levels each method has}

\item{pmf_pos, pmf_neg}{Used for ordinal methods. A n_method by n_level matrix representing the probability mass functions for positive and negative results, respectively}

\item{method_names}{Optional vector of names used to identify each method}

\item{level_names}{Used for ordinal methods. Optional vector of names used to identify each level}

\item{obs_names}{Optional vector of names used to identify each observation}
}
\value{
A list containing the simulated data set and the parameters used to generate it
}
\description{
Generate data for multiple ordinal measurement methods
}
