% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{getNames}
\alias{getNames}
\title{Return names from MultiMethodMLEstimate object}
\usage{
getNames(x, name)
}
\arguments{
\item{x}{An object of class MultiMethodMLEstimate.}

\item{name}{type of names to extract, e.g. "method_names", "obs_names", "level_names"}
}
\value{
Contents of \code{names} slot of the MultiMethodMLEstimate object.
}
\description{
Returns the names slot from a MultiMethodMLEstimate object
}
