% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{getResults,MultiMethodMLEstimate-method}
\alias{getResults,MultiMethodMLEstimate-method}
\title{Return result from MultiMethodMLEstimate object}
\usage{
\S4method{getResults}{MultiMethodMLEstimate}(x)
}
\arguments{
\item{x}{An object of class MultiMethodMLEstimate.}
}
\value{
Contents of \code{results} slot of the MultiMethodMLEstimate object.
}
\description{
Returns the result slot from a MultiMethodMLEstimate object
}
