% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\docType{class}
\name{MultiMethodMLEstimate-class}
\alias{MultiMethodMLEstimate-class}
\title{S4 object for storing the results of multi-method ML accuracy estimates}
\description{
S4 object for storing the results of multi-method ML accuracy estimates
}
\section{Slots}{

\describe{
\item{\code{results}}{a list of estimated accuracy statistics}

\item{\code{names}}{a list containing the names of various dimensions}

\item{\code{data}}{a copy of the data used to generate the estimate}

\item{\code{iter}}{an integer number of iterations used by EM algorithm}

\item{\code{prog}}{a list of interim statistic estimates}

\item{\code{type}}{a string describing the data type}
}}

