% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{name_thing}
\alias{name_thing}
\title{Create unique names for a set of things}
\usage{
name_thing(thing = "", n = 1)
}
\arguments{
\item{thing}{a string that describes the set of items to name}

\item{n}{an integer number of unique names to create}
}
\value{
a vector of unique names
}
\description{
Create unique names for a set of things
}
