% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{plot,MultiMethodMLEstimate-method}
\alias{plot,MultiMethodMLEstimate-method}
\title{Call \code{plot_ML} on MultiMethodMLEstimate S4 objects}
\usage{
\S4method{plot}{MultiMethodMLEstimate}(x, y, ...)
}
\arguments{
\item{x}{MultiMethodMLEstimate S4 object}

\item{y}{not used}

\item{...}{Additional arguments}
}
\value{
A list of ggplot2 plots
}
\description{
Call \code{plot_ML} on MultiMethodMLEstimate S4 objects
}
