% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{plot.boot_ML}
\alias{plot.boot_ML}
\title{Plot univariate distributions of bootstrapped ML estimates}
\usage{
\method{plot}{boot_ML}(x, probs = c(0.1, 0.5, 0.9), ...)
}
\arguments{
\item{x}{a result created by calling \code{boot_ML} on a \code{MultiMethodMLEstimate} object.}

\item{probs}{a vector of distribution quantile values to indicate with vertical lines.}

\item{...}{additional arguments.}
}
\value{
a named list of named plots.
}
\description{
\code{plot.boot_ML()} creates univariate plots of bootstrap results from \code{boot_ML()}.
}
