% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot_boot_ML}
\alias{plot_boot_ML}
\title{Plot univariate distributions of bootstrapped ML estimates}
\usage{
plot_boot_ML(boot_ML_result, probs = c(0.05, 0.5, 0.95))
}
\arguments{
\item{boot_ML_result}{a result created by calling \code{boot_ML} on a \code{MultiMethodMLEstimate} object.}

\item{probs}{distribution quantile values to indicate with vertical lines.}
}
\value{
a named list of named plots.
}
\description{
\code{plot_boot_ML()} creates univariate plots of bootstrap results from \code{boot_ML()}.
}
