% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot_ML}
\alias{plot_ML}
\title{Create plots of ML estimation process}
\usage{
plot_ML(ML_est, params = NULL)
}
\arguments{
\item{ML_est}{A MultiMethodEstimate class object}

\item{params}{An optional list of population parameters}

\item{type}{A string specifying the data type of the methods under evaluation}
}
\value{
a list of ggplot2 plots
}
\description{
Create plots of ML estimation process
}
