% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_method_functions.R
\name{plot_ML_binary}
\alias{plot_ML_binary}
\title{Generate plots from MultiMethodEstimate class objects}
\usage{
plot_ML_binary(
  ML_est,
  params = list(prev = NULL, se = NULL, sp = NULL, D = NULL)
)
}
\arguments{
\item{ML_est}{A MultiMethodEstimate class object}

\item{params}{A named list of population parameters, i.e. the "correct" values. This is used when the data are simulated and/or when the true values of the results statistics are known.}
}
\value{
A list of plots
}
\description{
Generate plots from MultiMethodEstimate class objects
}
