% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_method_functions.R
\name{plot_ML_ordinal}
\alias{plot_ML_ordinal}
\title{Create plots of ordinal multimethod comparison estimates}
\usage{
plot_ML_ordinal(
  ML_est,
  params = list(pi_1_1 = NULL, phi_1ij_1 = NULL, phi_0ij_1 = NULL, D = NULL, obs_names =
    NULL)
)
}
\arguments{
\item{ML_est}{A MultiMethodMLEstimate class object with type "Ordinal"}

\item{params}{A list of population parameters. This is mostly used to evaluate
results from a simulation (where parameters are already known), but can be used to visualize
results with respect to some True value.}
}
\value{
a list of ggplot2 plots
}
\description{
Create plots of ordinal multimethod comparison estimates
}
