% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_method_functions.R
\name{pollinate_ML_binary}
\alias{pollinate_ML_binary}
\title{Generate starting values for EM algorithm}
\usage{
pollinate_ML_binary(data)
}
\arguments{
\item{data}{A matrix with n_obs rows and n_method columns}
}
\value{
A named list containing the seed values for the EM algorithm
}
\description{
Generate starting values for EM algorithm
}
