% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_method_functions.R
\name{pollinate_ML_continuous}
\alias{pollinate_ML_continuous}
\title{Create initialization values for ML estimation}
\usage{
pollinate_ML_continuous(
  t_k,
  prev = 0.5,
  q_seeds = c((1 - prev)/2, 1 - (prev/2)),
  high_pos = TRUE
)
}
\arguments{
\item{t_k}{A data set of \code{n_obs} rows and \code{n_method} columns}

\item{prev}{A double between 0-1 representing the proportion of positives in the population}

\item{q_seeds}{A vector of length 2 representing the quantiles at which the two groups are assumed to be centered}

\item{high_pos}{A logical indicating whether larger values are considered "positive"}
}
\description{
Create initialization values for ML estimation
}
