% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_method_functions.R
\name{pollinate_ML_ordinal}
\alias{pollinate_ML_ordinal}
\title{Create Seed Values for EM Algorithm}
\usage{
pollinate_ML_ordinal(
  data,
  n_level = NULL,
  threshold_level = ceiling(n_level/2),
  level_names = NULL
)
}
\arguments{
\item{data}{a}

\item{n_level}{a}

\item{threshold_level}{a}

\item{level_names}{a}
}
\value{
A list of initial values for EM algorithm
}
\description{
Create Seed Values for EM Algorithm
}
