% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_method_functions.R
\name{random_start_binary}
\alias{random_start_binary}
\title{Initialize random starting values}
\usage{
random_start_binary(n_method = NULL, method_names = NULL)
}
\value{
List containing initial values to be passed to \code{init} argument
}
\description{
Creates random initial sensitivity and specificity values for \code{n_method} methods.
Values are generated from a random beta distribution with shape parameters
\code{a=3} and \code{b=1}. Prevalence is a random value from a uniform distribution.
}
