% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{setFreqs}
\alias{setFreqs}
\title{Set frequencies of MultiMethodMLEstimate objects}
\usage{
setFreqs(x, freqs = NULL)
}
\arguments{
\item{x}{An object of class MultiMethodMLEstimate.}

\item{freqs}{A vector of non-negative integers}
}
\value{
A copy of the MultiMethodMLEstimate object with \code{freqs} in the appropriate slot.
}
\description{
Set the freqs slot of a MultiMethodMLEstimate object. This can be used to update
objects created using \code{emery} versions < 0.6.1 to work with newer versions.
}
