% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{show,MultiMethodMLEstimate-method}
\alias{show,MultiMethodMLEstimate-method}
\title{Print final results stored in MultiMethodMLEstimate S4 object}
\usage{
\S4method{show}{MultiMethodMLEstimate}(object)
}
\arguments{
\item{object}{MultiMethodMLEstimate.}
}
\description{
Print final results stored in MultiMethodMLEstimate S4 object
}
