% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{unique_obs_summary}
\alias{unique_obs_summary}
\title{Reduce Data by Summarizing Observations}
\usage{
unique_obs_summary(data)
}
\arguments{
\item{data}{An \code{n_obs} by \code{n_method} \code{\link{matrix}} containing the
observed values for each method. If the dimensions are named, row names will
be used to name each observation (\code{obs_names}) and column names will be used
to name each measurement method (\code{method_names}).}
}
\value{
list containing matrix of unique rows with names and frequencies of duplicates.
}
\description{
Extract unique rows from a data frame. Collect row names of duplicate observations and frequencies of each.
}
