% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{define_disease_state}
\alias{define_disease_state}
\title{Define the True disease state of a simulated sample}
\usage{
define_disease_state(D = NULL, n_obs = NULL, prev = NULL)
}
\arguments{
\item{D}{Optional binary vector representing the true classification of
each observation.}

\item{n_obs}{An integer representing the number of observations to simulate.}

\item{prev}{A value between 0-1 which represents the proportion of
"positive" results in the target population.}
}
\value{
A list of features defining the true disease status of each observation
}
\description{
Define the True disease state of a simulated sample
}
