% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{estimate_ML}
\alias{estimate_ML}
\title{Estimate maximum likelihood accuracy statistics by expectation maximization}
\usage{
estimate_ML(
  type = c("binary", "ordinal", "continuous"),
  data,
  init = list(NULL),
  max_iter = 1000,
  tol = 1e-07,
  save_progress = TRUE,
  ...
)
}
\arguments{
\item{type}{A string specifying the data type of the methods under evaluation}

\item{data}{An n_obs by n_method matrix containing the observed values for each method. If dimensions are named, row names will be used for obs_names, and column names will be used for method_names}

\item{init}{An optional list of initial values used to seed the EM algorithm. If initial values are not provided, a function will be called on the data to estimate starting values.}

\item{max_iter}{The maximum number of EM algorithm iterations to compute before reporting a result.}

\item{tol}{The minimum change in statistic estimates needed to continue iterating the EM algorithm.}

\item{save_progress}{A logical indication whether to save interim calculations used in the EM algorithm.}

\item{...}{
  Arguments passed on to \code{\link[=estimate_ML_ordinal]{estimate_ML_ordinal}}
  \describe{
    \item{\code{level_names}}{An optional, ordered, character vector of unique names corresponding to
the levels of the methods.}
  }}
}
\value{
a MultiMethodMLEstimate class S4 object containing ML accuracy statistics by EM
}
\description{
Estimate maximum likelihood accuracy statistics by expectation maximization
}
