% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_method_functions.R
\name{estimate_ML_binary}
\alias{estimate_ML_binary}
\title{Estimate accuracy statistics and prevalence by ML}
\usage{
estimate_ML_binary(
  data,
  init = list(prev_1 = NULL, se_1 = NULL, sp_1 = NULL),
  max_iter = 100,
  tol = 1e-07,
  save_progress = TRUE
)
}
\arguments{
\item{data}{An n_obs by n_method matrix containing the observed values for each method. If dimensions are named, row names will be used for obs_names, and column names will be used for method_names}

\item{init}{An optional list of initial values used to seed the EM algorithm. If initial values are not provided, a function will be called on the data to estimate starting values.}

\item{max_iter}{The maximum number of EM algorithm iterations to compute before reporting a result.}

\item{tol}{The minimum change in statistic estimates needed to continue iterating the EM algorithm.}

\item{save_progress}{A logical indication whether to save interim calculations used in the EM algorithm.}
}
\value{
A list of statistics resulting from the application of the EM algorithm.
If save_progress is TRUE, a list of each value needed by the algorithm for each
iteration will also be included.
}
\description{
Estimate accuracy statistics and prevalence by ML
}
