% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{plot,MultiMethodMLEstimate-method}
\alias{plot,MultiMethodMLEstimate-method}
\title{Plot Results Stored in MultiMethodMLEstimate S4 Objects}
\usage{
\S4method{plot}{MultiMethodMLEstimate}(x, y, ...)
}
\arguments{
\item{x}{MultiMethodMLEstimate.}
}
\value{
a list of ggplot2 plots
}
\description{
Plot Results Stored in MultiMethodMLEstimate S4 Objects
}
