% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{plot,MultiMethodMLEstimate-method}
\alias{plot,MultiMethodMLEstimate-method}
\title{Create plots from a MultiMethodMLEstimate object}
\usage{
\S4method{plot}{MultiMethodMLEstimate}(x, y, ...)
}
\arguments{
\item{x}{a MultiMethodMLEstimate S4 object}

\item{y}{not used}

\item{...}{
  Arguments passed on to \code{\link[=plot_ML]{plot_ML}}
  \describe{
    \item{\code{params}}{A list of population parameters. This is primarily used to evaluate
results from a simulation where the target parameters are known, but can be used to visualize
results with respect to some True value.}
  }}
}
\value{
A list of ggplot2 plots
}
\description{
Create a list of plots visualizing the expectation maximization process and
resulting accuracy statistics stored in a MultiMethodMLEstimate object.
}
