% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R,
%   R/binary_method_functions.R, R/ordinal_method_functions.R,
%   R/continuous_method_functions.R
\name{plot_ML}
\alias{plot_ML}
\alias{plot_ML_binary}
\alias{plot_ML_ordinal}
\alias{plot_ML_continuous}
\title{Create plots visualizing the ML estimation process.}
\usage{
plot_ML(ML_est, params = NULL)

plot_ML_binary(
  ML_est,
  params = list(prev = NULL, se = NULL, sp = NULL, D = NULL)
)

plot_ML_ordinal(
  ML_est,
  params = list(pi_1_1 = NULL, phi_1ij_1 = NULL, phi_0ij_1 = NULL, D = NULL)
)

plot_ML_continuous(
  ML_est,
  params = list(prev_1 = NULL, mu_i1_1 = NULL, sigma_i1_1 = NULL, mu_i0_1 = NULL,
    sigma_i0_1 = NULL, D = NULL)
)
}
\arguments{
\item{ML_est}{A MultiMethodMLEstimate class object}

\item{params}{A list of population parameters. This is mostly used to evaluate
results from a simulation where the target parameters are already known, but can be used to visualize
results with respect to some True value.}
}
\value{
a list of ggplot2 plots.
}
\description{
Create plots visualizing the ML estimation process.
}
\examples{
# Set seed for this example
set.seed(11001101)

# Generate data for 4 binary methods
my_sim <- generate_multimethod_data(
  "binary",
  n_obs = 75,
  n_method = 4,
  se = c(0.87, 0.92, 0.79, 0.95),
  sp = c(0.85, 0.93, 0.94, 0.80),
  method_names = c("alpha", "beta", "gamma", "delta"))

# View the data
my_sim$generated_data

# View the parameters used to generate the data
my_sim$params

# Estimate ML accuracy values by EM algorithm
my_result <- estimate_ML(
  "binary",
  data = my_sim$generated_data,
  save_progress = FALSE # this reduces the data stored in the resulting object
)

# View results of ML estimate
my_result@results

}
