% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_method_functions.R
\name{pollinate_ML_binary}
\alias{pollinate_ML_binary}
\title{Create Seed Values for EM Algorithm}
\usage{
pollinate_ML_binary(data)
}
\arguments{
\item{data}{A matrix with n_obs rows and n_method columns}
}
\value{
A named list containing seed values for the EM algorithm
}
\description{
Create Seed Values for EM Algorithm
}
