% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_method_functions.R
\name{pollinate_ML_ordinal}
\alias{pollinate_ML_ordinal}
\title{Title}
\usage{
pollinate_ML_ordinal(
  data,
  n_level = NULL,
  threshold_level = ceiling(n_level/2),
  level_names = NULL
)
}
\arguments{
\item{level_names}{}
}
\description{
Title
}
